New file send method
