+++ FileteleporterTransfert/Server/ServerHandle.cs
@@ -48,18 +48,6 @@ namespace server
                 writer.Write(file);
             }
             Console.WriteLine("finished");
-
-            //public static void PlayerMovement(int _fromClient, Packet _packet)
-            //{
-            //    bool[] _inputs = new bool[_packet.ReadInt()];
-            //    for (int i = 0; i < _inputs.Length; i++)
-            //    {
-            //        _inputs[i] = _packet.ReadBool();
-            //    }
-            //    Quaternion _rotation = _packet.ReadQuaternion();
-
-            //    Server.clients[_fromClient].player.SetInput(_inputs, _rotation);
-            //}
         }
     }
 }
